+++ packages/mobile/locales/es-419/fiatExchangeFlow.json
@@ -1,6 +1,6 @@
 {
-  "add_funds": "Añadir fondos",
-  "cash_out": "Retirar",
+  "addFunds": "Añadir fondos",
+  "cashOut": "Retirar",
   "celoDollarsAt": "Celo Dollars @ <1></1>",
   "dislamerCeloDollars": "Los valores del dólar celo son estimados.",
   "pleaseSelectProvider": "Seleccione un proveedor a continuación. Estos enlaces se conectan a un servicio de terceros.",
