+++ packages/mobile/src/exchange/ExchangeTradeScreen.tsx
@@ -73,6 +73,8 @@ const mapStateToProps = (state: RootState): StateProps => ({
   localCurrencyExchangeRate: getLocalCurrencyExchangeRate(state),
 })
 
+const safeAreaInset = { top: 'never' as 'never', bottom: 'always' as 'always' }
+
 export class ExchangeTradeScreen extends React.Component<Props, State> {
   state: State = {
     inputToken: CURRENCY_ENUM.GOLD,
@@ -281,7 +283,7 @@ export class ExchangeTradeScreen extends React.Component<Props, State> {
       <SafeAreaView
         // Force inset as this screen uses auto focus and KeyboardSpacer padding is initially
         // incorrect because of that
-        forceInset={{ top: 'never', bottom: 'always' }}
+        forceInset={safeAreaInset}
         style={styles.container}
       >
         <DisconnectBanner />
