+++ packages/mobile/src/fiatExchanges/FiatExchangeOptions.tsx
@@ -1,13 +1,23 @@
+import colors from '@celo/react-components/styles/colors.v2'
 import fontStyles from '@celo/react-components/styles/fonts.v2'
 import variables from '@celo/react-components/styles/variables'
+import { NavigationProp, RouteProp } from '@react-navigation/core'
 import { StackScreenProps } from '@react-navigation/stack'
 import * as React from 'react'
 import { useTranslation } from 'react-i18next'
 import { Image, SafeAreaView, ScrollView, StyleSheet, Text, View } from 'react-native'
+import { CustomEventNames } from 'src/analytics/constants'
+import BackButton from 'src/components/BackButton.v2'
+import CurrencyDisplay from 'src/components/CurrencyDisplay'
 import Dialog from 'src/components/Dialog'
 import ListItem from 'src/fiatExchanges/ListItem'
+import { CURRENCIES, CURRENCY_ENUM } from 'src/geth/consts'
+import i18n from 'src/i18n'
+import QuestionIcon from 'src/icons/QuestionIcon'
 import { moonpayLogo } from 'src/images/Images'
+import { emptyHeader, HeaderTitleWithSubtitle } from 'src/navigator/Headers.v2'
 import { Screens } from 'src/navigator/Screens'
+import { TopBarIconButton } from 'src/navigator/TopBarButton.v2'
 import { StackParamList } from 'src/navigator/types'
 
 type RouteProps = StackScreenProps<StackParamList, Screens.FiatExchangeOptions>
@@ -16,6 +26,37 @@ type Props = RouteProps
 interface Provider {
   image: React.ReactNode
   screen: keyof StackParamList
+}
+
+export const fiatExchangesOptionsScreenOptions = ({
+  route,
+  navigation,
+}: {
+  route: RouteProp<StackParamList, Screens.FiatExchangeOptions>
+  navigation: NavigationProp<StackParamList, Screens.FiatExchangeOptions>
+}) => {
+  function showExplanation() {
+    navigation.setParams({ isExplanationOpen: true })
+  }
+  const amount = (
+    <CurrencyDisplay
+      amount={{ value: route.params.amount, currencyCode: CURRENCIES[CURRENCY_ENUM.DOLLAR].code }}
+    />
+  )
+  return {
+    ...emptyHeader,
+    headerLeft: () => <BackButton eventName={CustomEventNames.send_amount_back} />,
+    headerTitle: () => (
+      <HeaderTitleWithSubtitle
+        title={amount}
+        subTitle={i18n.t(`fiatExchangeFlow:${route.params?.isAddFunds ? 'addFunds' : 'cashOut'}`)}
+      />
+    ),
+    headerRight: () => (
+      <TopBarIconButton icon={<QuestionIcon color={colors.greenUI} />} onPress={showExplanation} />
+    ),
+    headerRightContainerStyle: { paddingRight: 16 },
+  }
 }
 
 function FiatExchangeOptions({ route, navigation }: Props) {
