+++ packages/mobile/src/fiatExchanges/MoonPay.tsx
@@ -5,11 +5,27 @@ import { WebView } from 'react-native-webview'
 import { useSelector } from 'react-redux'
 import { showError } from 'src/alert/actions'
 import { ErrorMessages } from 'src/app/ErrorMessages'
-import { SIGN_MOONPAY_FIREBASE_URL } from 'src/config'
+import config from 'src/geth/networkConfig'
+import i18n from 'src/i18n'
+import { emptyHeader } from 'src/navigator/Headers.v2'
+import { navigate } from 'src/navigator/NavigationService'
+
 import { getLocalCurrencyCode } from 'src/localCurrency/selectors'
+import { Screens } from 'src/navigator/Screens'
+import { TopBarTextButton } from 'src/navigator/TopBarButton.v2'
 import { currentAccountSelector } from 'src/web3/selectors'
 
 const celoCurrencyCode = 'CUSD'
+
+export const moonPayOptions = () => {
+  const navigateToFiatExchange = () => navigate(Screens.FiatExchange)
+  return {
+    ...emptyHeader,
+    headerLeft: () => (
+      <TopBarTextButton title={i18n.t('global:done')} onPress={navigateToFiatExchange} />
+    ),
+  }
+}
 
 function FiatExchangeWeb() {
   const [uri, setUri] = React.useState('')
@@ -17,7 +33,7 @@ function FiatExchangeWeb() {
   const localCurrencyCode = useSelector(getLocalCurrencyCode)
   React.useEffect(() => {
     const getSignedUrl = async () => {
-      const response = await fetch(SIGN_MOONPAY_FIREBASE_URL, {
+      const response = await fetch(config.signMoonpayUrl, {
         method: 'POST',
         headers: {
           Accept: 'application/json',
