+++ packages/mobile/src/geth/networkConfig.ts
@@ -19,6 +19,7 @@ interface NetworkConfig {
   blockchainApiUrl: string
   pgpnpUrl: string // Phone Number Privacy service url
   pgpnpPubKey: string
+  signMoonpayUrl: string
 }
 
 const pgpnpUrlStaging = 'https://us-central1-celo-phone-number-privacy-stg.cloudfunctions.net'
@@ -27,6 +28,7 @@ const pgpnpPubKeyStaging =
 const pgpnpUrlProd = 'https://us-central1-celo-phone-number-privacy.cloudfunctions.net'
 const pgpnpPubKeyProd =
   '6VFX/ufxvL54NDRlJMe0jlbb9wD3L/Kfm6K2qEKoxrDMS42Q1S7ZOH88tdpOJhcAJPdDTjGE6qHQBiV48n/jctSuOJ8HjsRzp6VJWMHW3imqep7nwyhzNfLcPXJfrngB'
+const signMoonpayUrlStaging = 'https://us-central1-celo-org-mobile.cloudfunctions.net/signMoonpay'
 
 const networkConfigs: { [testnet: string]: NetworkConfig } = {
   [Testnets.integration]: {
@@ -36,6 +38,7 @@ const networkConfigs: { [testnet: string]: NetworkConfig } = {
     blockchainApiUrl: 'https://integration-dot-celo-testnet.appspot.com/',
     pgpnpUrl: pgpnpUrlStaging,
     pgpnpPubKey: pgpnpPubKeyStaging,
+    signMoonpayUrl: signMoonpayUrlStaging,
   },
   [Testnets.alfajoresstaging]: {
     nodeDir: `.${Testnets.alfajoresstaging}`,
@@ -44,6 +47,7 @@ const networkConfigs: { [testnet: string]: NetworkConfig } = {
     blockchainApiUrl: 'https://alfajoresstaging-dot-celo-testnet.appspot.com/',
     pgpnpUrl: pgpnpUrlStaging,
     pgpnpPubKey: pgpnpPubKeyStaging,
+    signMoonpayUrl: signMoonpayUrlStaging,
   },
   [Testnets.alfajores]: {
     nodeDir: `.${Testnets.alfajores}`,
@@ -52,6 +56,7 @@ const networkConfigs: { [testnet: string]: NetworkConfig } = {
     blockchainApiUrl: 'https://alfajores-dot-celo-testnet-production.appspot.com/',
     pgpnpUrl: pgpnpUrlProd,
     pgpnpPubKey: pgpnpPubKeyProd,
+    signMoonpayUrl: signMoonpayUrlStaging,
   },
   [Testnets.pilot]: {
     nodeDir: `.${Testnets.pilot}`,
@@ -60,6 +65,7 @@ const networkConfigs: { [testnet: string]: NetworkConfig } = {
     blockchainApiUrl: 'https://pilot-dot-celo-testnet-production.appspot.com/',
     pgpnpUrl: pgpnpUrlStaging,
     pgpnpPubKey: pgpnpPubKeyStaging,
+    signMoonpayUrl: signMoonpayUrlStaging,
   },
   [Testnets.pilotstaging]: {
     nodeDir: `.${Testnets.pilotstaging}`,
@@ -68,6 +74,7 @@ const networkConfigs: { [testnet: string]: NetworkConfig } = {
     blockchainApiUrl: 'https://pilotstaging-dot-celo-testnet.appspot.com/',
     pgpnpUrl: pgpnpUrlStaging,
     pgpnpPubKey: pgpnpPubKeyStaging,
+    signMoonpayUrl: signMoonpayUrlStaging,
   },
   [Testnets.baklavastaging]: {
     nodeDir: `.${Testnets.baklavastaging}`,
@@ -76,6 +83,7 @@ const networkConfigs: { [testnet: string]: NetworkConfig } = {
     blockchainApiUrl: 'https://baklavastaging-dot-celo-testnet.appspot.com/',
     pgpnpUrl: pgpnpUrlStaging,
     pgpnpPubKey: pgpnpPubKeyStaging,
+    signMoonpayUrl: signMoonpayUrlStaging,
   },
   [Testnets.baklava]: {
     nodeDir: `.${Testnets.baklava}`,
@@ -84,6 +92,7 @@ const networkConfigs: { [testnet: string]: NetworkConfig } = {
     blockchainApiUrl: 'https://baklava-dot-celo-testnet-production.appspot.com/',
     pgpnpUrl: pgpnpUrlStaging,
     pgpnpPubKey: pgpnpPubKeyStaging,
+    signMoonpayUrl: signMoonpayUrlStaging,
   },
 }
