+++ packages/mobile/src/navigator/Navigator.tsx
@@ -1,5 +1,5 @@
 import colors from '@celo/react-components/styles/colors.v2'
-import { NavigationProp, RouteProp } from '@react-navigation/core'
+import { RouteProp } from '@react-navigation/core'
 import { createStackNavigator } from '@react-navigation/stack'
 import * as React from 'react'
 import SplashScreen from 'react-native-splash-screen'
@@ -30,7 +30,6 @@ import BackupSocial from 'src/backup/BackupSocial'
 import BackupSocialIntro from 'src/backup/BackupSocialIntro'
 import BackButton from 'src/components/BackButton.v2'
 import CancelButton from 'src/components/CancelButton.v2'
-import CurrencyDisplay from 'src/components/CurrencyDisplay'
 import DappKitAccountScreen from 'src/dappkit/DappKitAccountScreen'
 import DappKitSignTxScreen from 'src/dappkit/DappKitSignTxScreen'
 import DappKitTxDataScreen from 'src/dappkit/DappKitTxDataScreen'
@@ -39,12 +38,15 @@ import ReclaimPaymentConfirmationScreen from 'src/escrow/ReclaimPaymentConfirmat
 import ExchangeReview from 'src/exchange/ExchangeReview'
 import ExchangeTradeScreen from 'src/exchange/ExchangeTradeScreen'
 import FeeExchangeEducation from 'src/exchange/FeeExchangeEducation'
-import FiatExchangeAmount from 'src/fiatExchanges/FiatExchangeAmount'
-import FiatExchangeOptions from 'src/fiatExchanges/FiatExchangeOptions'
-import MoonPay from 'src/fiatExchanges/MoonPay'
-import { CURRENCIES, CURRENCY_ENUM } from 'src/geth/consts'
+import FiatExchangeAmount, {
+  fiatExchangesAmountScreenOptions,
+} from 'src/fiatExchanges/FiatExchangeAmount'
+import FiatExchangeOptions, {
+  fiatExchangesOptionsScreenOptions,
+} from 'src/fiatExchanges/FiatExchangeOptions'
+import MoonPay, { moonPayOptions } from 'src/fiatExchanges/MoonPay'
+import { CURRENCY_ENUM } from 'src/geth/consts'
 import i18n from 'src/i18n'
-import QuestionIcon from 'src/icons/QuestionIcon'
 import PhoneNumberLookupQuotaScreen from 'src/identity/PhoneNumberLookupQuotaScreen'
 import ImportWallet from 'src/import/ImportWallet'
 import ImportWalletEmpty from 'src/import/ImportWalletEmpty'
@@ -64,10 +66,10 @@ import {
   nuxNavigationOptions,
   nuxNavigationOptionsNoBackButton,
 } from 'src/navigator/Headers.v2'
-import { navigate, navigateBack, navigateToExchangeHome } from 'src/navigator/NavigationService'
+import { navigateBack, navigateToExchangeHome } from 'src/navigator/NavigationService'
 import QRNavigator from 'src/navigator/QRNavigator'
 import { Screens } from 'src/navigator/Screens'
-import { TopBarIconButton, TopBarTextButton } from 'src/navigator/TopBarButton.v2'
+import { TopBarTextButton } from 'src/navigator/TopBarButton.v2'
 import { StackParamList } from 'src/navigator/types'
 import ImportContactsScreen from 'src/onboarding/contacts/ImportContactsScreen'
 import OnboardingSuccessScreen from 'src/onboarding/success/OnboardingSuccessScreen'
@@ -362,64 +364,6 @@ const backupScreens = (Navigator: typeof Stack) => (
   </>
 )
 
-const fiatExchangesAmountScreenOptions = ({
-  route,
-}: {
-  route: RouteProp<StackParamList, Screens.FiatExchangeAmount>
-}) => {
-  return {
-    ...emptyHeader,
-    headerLeft: () => <BackButton eventName={CustomEventNames.send_amount_back} />,
-    headerTitle: () => (
-      <HeaderTitleWithBalance
-        title={i18n.t(`fiatExchangeFlow:${route.params?.isAddFunds ? 'add_funds' : 'cash_out'}`)}
-        token={CURRENCY_ENUM.DOLLAR}
-      />
-    ),
-  }
-}
-
-const fiatExchangesOptionsScreenOptions = ({
-  route,
-  navigation,
-}: {
-  route: RouteProp<StackParamList, Screens.FiatExchangeOptions>
-  navigation: NavigationProp<StackParamList, Screens.FiatExchangeOptions>
-}) => {
-  function showExplanation() {
-    navigation.setParams({ isExplanationOpen: true })
-  }
-  const amount = (
-    <CurrencyDisplay
-      amount={{ value: route.params.amount, currencyCode: CURRENCIES[CURRENCY_ENUM.DOLLAR].code }}
-    />
-  )
-  return {
-    ...emptyHeader,
-    headerLeft: () => <BackButton eventName={CustomEventNames.send_amount_back} />,
-    headerTitle: () => (
-      <HeaderTitleWithSubtitle
-        title={amount}
-        subTitle={i18n.t(`fiatExchangeFlow:${route.params?.isAddFunds ? 'add_funds' : 'cash_out'}`)}
-      />
-    ),
-    headerRight: () => (
-      <TopBarIconButton icon={<QuestionIcon color={colors.greenUI} />} onPress={showExplanation} />
-    ),
-    headerRightContainerStyle: { paddingRight: 16 },
-  }
-}
-
-const fiatExchangesProviderWebViewScreenOptions = () => {
-  const navigateToFiatExchange = () => navigate(Screens.FiatExchange)
-  return {
-    ...emptyHeader,
-    headerLeft: () => (
-      <TopBarTextButton title={i18n.t('global:done')} onPress={navigateToFiatExchange} />
-    ),
-  }
-}
-
 const settingsScreens = (Navigator: typeof Stack) => (
   <>
     <Navigator.Screen options={noHeader} name={Screens.Account} component={Account} />
@@ -463,11 +407,7 @@ const settingsScreens = (Navigator: typeof Stack) => (
       name={Screens.FiatExchangeOptions}
       component={FiatExchangeOptions}
     />
-    <Navigator.Screen
-      options={fiatExchangesProviderWebViewScreenOptions}
-      name={Screens.MoonPay}
-      component={MoonPay}
-    />
+    <Navigator.Screen options={moonPayOptions} name={Screens.MoonPay} component={MoonPay} />
   </>
 )
