--- packages/mobile/src/stableToken/selectors.ts
@@ -1,2 +1,0 @@
-import { RootState } from 'src/redux/reducers'
-export const dollarBalanceSelector = (state: RootState) => state.stableToken.balance
